requirements.txt
streamlit==1.32.0
sentence-transformers==2.5.1
faiss-cpu==1.8.0.post1
pandas==2.1.4
numpy==1.26.4
langchain==0.1.0
langchain-ibm==0.1.0
langchain-community==0.0.20
langsmith==0.0.87
python-dotenv==1.0.0
requests==2.31.0
plotly==5.18.0
pypdf==3.17.0
SpeechRecognition==3.10.4
pyttsx3==2.90
gtts==2.5.1
audio-recorder-streamlit==0.0.8
